Types

PeriodicSchurDecompositions.PeriodicSchur — Type

PeriodicSchur

Matrix factorization type of the periodic Schur factorization of a series A₁, A₂, ... Aₚ of matrices. This is the return type of pschur!(_).

The orientation property may be 'L'(left), corresponding to the product Aₚ * Aₚ₋₁ * ... * A₂ * A₁ or 'R'(right), for the product A₁ * A₂ * ... * Aₚ.

The decomposition for the "right" orientation is Z₁' * A₁ * Z₂ = T₁; Z₂' * A₂ * Z₃ = T₂; ...; Zₚ' * Aₚ * Z₁ = Tₚ.

The decomposition for the "left" orientation is Z₂' * A₁ * Z₁ = T₁; Z₃' * A₂ * Z₂ = T₂; ...; Z₁' * Aₚ * Zₚ = Tₚ.

For real element types, Tₖ is a quasi-triangular "real Schur" matrix, where k is the value of the schurindex field. Otherwise the Tⱼ are upper triangular. The Zⱼ are unitary (orthogonal for reals).

Given F::PeriodicSchur, the (quasi) triangular Schur factor Tₖ can be obtained via F.T1. F.T is a vector of the remaining triangular Tⱼ. F.Z is a vector of the Zⱼ. F.values is a vector of the eigenvalues of the product of the Aⱼ.

PeriodicSchurDecompositions.GeneralizedPeriodicSchur — Type

GeneralizedPeriodicSchur

Matrix factorization type of the generalized periodic Schur factorization of a series A₁, A₂, ... Aₚ of matrices. This is the return type of pschur!(_) with a sign vector.

The orientation property may be 'L'(left), corresponding to the product Aₚ^sₚ * Aₚ₋₁^sₚ₋₁ * ... * A₂^s₂ * A₁^s₁ or 'R'(right), for the product A₁^s₁ * A₂^s₂ * ... * Aₚ^sₚ, where the signs sⱼ are '±1`.

The decomposition for the "right" orientation is Zⱼ' * Aⱼ * Zᵢ = Tⱼ where i=mod(j,p)+1 if sⱼ=1, and Zᵢ' * Aⱼ * Zⱼ = Tⱼ if sⱼ=-1.

The decomposition for the "left" orientation is Zᵢ' * Aⱼ * Zⱼ = Tⱼ where i=mod(j,p)+1 if sⱼ=1, and Zⱼ' * Aⱼ * Zᵢ = Tⱼ if sⱼ=-1.

For real element types, Tₖ is a quasi-triangular "real Schur" matrix, where k is the value of the schurindex field. Otherwise the Tⱼ are upper triangular. The Zⱼ are unitary (orthogonal for reals).

Given F::GeneralizedPeriodicSchur, the (quasi) triangular Schur factor Tₖ can be obtained via F.T1. F.T is a vector of the remaining triangular Tⱼ. F.Z is a vector of the Zⱼ. F.values is a vector of the eigenvalues of the product of the Aⱼ. (The eigenvalues are stored internally in scaled form to avoid over/underflow.)

PeriodicSchurDecompositions.PartialPeriodicSchur — Type

PartialPeriodicSchur

A partial periodic Schur decomposition of a series of matrices, with k Schur vectors of length n in the Zⱼ where typically k ≪ n.

The decomposition for the "left" orientation is A₁ * Z₁ = Z₂ * T₁; A₂ * Z₂ = Z₃ * T₂; ...; Aₚ * Zₚ = Z₁ * Tₚ.

Properties are similar to PeriodicSchur.