Functions

Periodic Schur decompositions

PeriodicSchurDecompositions.pschur — Function

pschur(A::Vector{S<:StridedMatrix}, lr::Symbol) -> F::PeriodicSchur

Computes a periodic Schur decomposition of a series of general square matrices with left (lr=:L) or right (lr=:R) product ordering.

Optional arguments wantT and wantZ, defaulting to true, are booleans which may be used to save time and memory by suppressing computation of the T and Z matrices. See PeriodicSchur for the resulting structure.

PeriodicSchurDecompositions.pschur! — Function

pschur!(A::Vector{S<:StridedMatrix}) -> F::PeriodicSchur

Same as pschur but uses the input matrices A as workspace.

pschur!(H1,Hs) -> PeriodicSchur

Computes a periodic Schur decomposition of a series of Hessenberg/triangular matrices.

If not wantT then just compute the eigenvalues. If wantZ then also compute orthogonal transformation matrices: if Q is unset, these will be the Zⱼ; if Q is set to a vector of matrices Qⱼ, they will be the products QⱼZⱼ.

pschur!(A::Vector{S<:StridedMatrix}, S::Vector{Bool}, lr::Symbol) -> F::GeneralizedPeriodicSchur

Computes a generalized periodic Schur decomposition of a series of general square matrices with left (lr=:L) or right (lr=:R) orientation.

Optional arguments wantT and wantZ, defaulting to true, are booleans which may be used to save time and memory by suppressing computation of the T and Z matrices. See GeneralizedPeriodicSchur for the resulting structure.

Currently S[1] must be true.

Missing docstring.

Missing docstring for gpschur. Check Documenter's build log for details.

Periodic Hessenberg decompositions

PeriodicSchurDecompositions.phessenberg! — Function

phessenberg!(A::Vector{<:AbstractMatrix}) -> (Hessenberg, Vector{QR})

reduce a series of p matrices A = [A₁ A₂ ... Aₚ] to upper Hessenberg/triangular form via a cycle of orthogonal similarity transformations

Q₁'A₁Q₂ = H₁ Q₂'A₂Q₃ = H₂ Qₚ'AₚQ₁ = Hₚ

H₁ is upper Hessenberg, the other Hⱼ are upper triangular.

Types

PeriodicSchurDecompositions.PeriodicSchur — Type

PeriodicSchur

Matrix factorization type of the periodic Schur factorization of a series A₁, A₂, ... Aₚ of matrices. This is the return type of pschur!(_).

The orientation property may be 'L'(left), corresponding to the product Aₚ * Aₚ₋₁ * ... * A₂ * A₁ or 'R'(right), for the product A₁ * A₂ * ... * Aₚ.

The decomposition for the "right" orientation is Z₁' * A₁ * Z₂ = T₁; Z₂' * A₂ * Z₃ = T₂; ...; Zₚ' * Aₚ * Z₁ = Tₚ.

The decomposition for the "left" orientation is Z₂' * A₁ * Z₁ = T₁; Z₃' * A₂ * Z₂ = T₂; ...; Z₁' * Aₚ * Zₚ = Tₚ.

For real element types, Tₖ is a quasi-triangular "real Schur" matrix, where k is the value of the schurindex field. Otherwise the Tⱼ are upper triangular. The Zⱼ are unitary (orthogonal for reals).

Given F::PeriodicSchur, the (quasi) triangular Schur factor Tₖ can be obtained via F.T1. F.T is a vector of the remaining triangular Tⱼ. F.Z is a vector of the Zⱼ. F.values is a vector of the eigenvalues of the product of the Aⱼ.