Functions

Periodic Schur decompositions

PeriodicSchurDecompositions.pschur — Function

pschur(A::Vector{S<:StridedMatrix}, lr::Symbol) -> F::PeriodicSchur

Computes a periodic Schur decomposition of a series of general square matrices with left (lr=:L) or right (lr=:R) product ordering.

Optional arguments wantT and wantZ, defaulting to true, are booleans which may be used to save time and memory by suppressing computation of the T and Z matrices. See PeriodicSchur for the resulting structure.

PeriodicSchurDecompositions.pschur! — Function

pschur!(A::Vector{S<:StridedMatrix}, lr::Symbol=:R) -> F::PeriodicSchur

Same as pschur but uses the input matrices A as workspace.

pschur!(H1,Hs) -> PeriodicSchur

Computes a periodic Schur decomposition of a series of Hessenberg/triangular matrices.

If not wantT then just compute the eigenvalues. If wantZ then also compute orthogonal transformation matrices: if Q is unset, these will be the Zⱼ; if Q is set to a vector of matrices Qⱼ, they will be the products QⱼZⱼ.

pschur!(A::Vector{<:StridedMatrix}, S::Vector{Bool}, lr::Symbol) -> F::GeneralizedPeriodicSchur

Computes a generalized periodic Schur decomposition of a series of general square matrices with left (lr=:L) or right (lr=:R) orientation.

Entries in S must be false for the corresponding entries in A which are effectively inverted, and true for the rest. Currently Sⱼ must be true for the leftmost term.

Optional arguments wantT and wantZ, defaulting to true, are booleans which may be used to save time and memory by suppressing computation of the T and Z matrices. See GeneralizedPeriodicSchur for the resulting structure. (S[j]=false corresponds to sⱼ=-1 in the decomposition shown there.)

PeriodicSchurDecompositions.gpschur — Function

gpschur(As, Bs) -> F::GeneralizedPeriodicSchur

Computes a generalized periodic Schur decomposition corresponding to the formal product Bₚ⁻¹Aₚ...B₁⁻¹A₁ of paired series of matrices in left operator order [A₁,...,Aₚ],[B₁,...,Bₚ].

Terms in the decomposition are actually shifted by one; this does not change the eigenvalues but requires attention when dealing with invariant subspaces.

LinearAlgebra.ordschur! — Method

ordschur!(P::AbstractPeriodicSchur{T}, select::AbstractVector{Bool})

reorder a periodic Schur decomposition so that the eigenvalues corresponding to true entries in select and their associated subspace are moved to the top.

Periodic Hessenberg decompositions

PeriodicSchurDecompositions.phessenberg! — Function

phessenberg!(A::Vector{<:AbstractMatrix}) -> (Hessenberg, Vector{QR})

reduce a series of p matrices A = [A₁ A₂ ... Aₚ] to upper Hessenberg/triangular form via a cycle of orthogonal similarity transformations

Q₁'A₁Q₂ = H₁ Q₂'A₂Q₃ = H₂ Qₚ'AₚQ₁ = Hₚ

H₁ is upper Hessenberg, the other Hⱼ are upper triangular.

Types

PeriodicSchurDecompositions.PeriodicSchur — Type

PeriodicSchur

Matrix factorization type of the periodic Schur factorization of a series A₁, A₂, ... Aₚ of matrices. This is the return type of pschur!(_).

The orientation property may be 'L'(left), corresponding to the product Aₚ * Aₚ₋₁ * ... * A₂ * A₁ or 'R'(right), for the product A₁ * A₂ * ... * Aₚ.

The decomposition for the "right" orientation is Z₁' * A₁ * Z₂ = T₁; Z₂' * A₂ * Z₃ = T₂; ...; Zₚ' * Aₚ * Z₁ = Tₚ.

The decomposition for the "left" orientation is Z₂' * A₁ * Z₁ = T₁; Z₃' * A₂ * Z₂ = T₂; ...; Z₁' * Aₚ * Zₚ = Tₚ.

For real element types, Tₖ is a quasi-triangular "real Schur" matrix, where k is the value of the schurindex field. Otherwise the Tⱼ are upper triangular. The Zⱼ are unitary (orthogonal for reals).

Given F::PeriodicSchur, the (quasi) triangular Schur factor Tₖ can be obtained via F.T1. F.T is a vector of the remaining triangular Tⱼ. F.Z is a vector of the Zⱼ. F.values is a vector of the eigenvalues of the product of the Aⱼ.

PeriodicSchurDecompositions.GeneralizedPeriodicSchur — Type

GeneralizedPeriodicSchur

Matrix factorization type of the generalized periodic Schur factorization of a series A₁, A₂, ... Aₚ of matrices. This is the return type of pschur!(_) with a sign vector.

The orientation property may be 'L'(left), corresponding to the product Aₚ^sₚ * Aₚ₋₁^sₚ₋₁ * ... * A₂^s₂ * A₁^s₁ or 'R'(right), for the product A₁^s₁ * A₂^s₂ * ... * Aₚ^sₚ, where the signs sⱼ are '±1`.

The decomposition for the "right" orientation is Zⱼ' * Aⱼ * Zᵢ = Tⱼ where i=mod(j,p)+1 if sⱼ=1, and Zᵢ' * Aⱼ * Zⱼ = Tⱼ if sⱼ=-1.

The decomposition for the "left" orientation is Zᵢ' * Aⱼ * Zⱼ = Tⱼ where i=mod(j,p)+1 if sⱼ=1, and Zⱼ' * Aⱼ * Zᵢ = Tⱼ if sⱼ=-1.

For real element types, Tₖ is a quasi-triangular "real Schur" matrix, where k is the value of the schurindex field. Otherwise the Tⱼ are upper triangular. The Zⱼ are unitary (orthogonal for reals).

Given F::GeneralizedPeriodicSchur, the (quasi) triangular Schur factor Tₖ can be obtained via F.T1. F.T is a vector of the remaining triangular Tⱼ. F.Z is a vector of the Zⱼ. F.values is a vector of the eigenvalues of the product of the Aⱼ. (The eigenvalues are stored internally in scaled form to avoid over/underflow.)